loader_docker/src/requirements.txt
pypardot~=0.1
smart_open~=5.1
boto3~=1.17
pandas~=1.2
requests~=2.23
snowflake-connector-python~=2.4.6